Refactor API Errors
