+++ app/api_v1/__init__.py
@@ -4,4 +4,3 @@ api = Blueprint('api_v1', __name__)
 from app.api_v1 import resources
 from app.api_v1 import errors
 from app.api_v1 import index
-from app.api_v1 import fhir_errors
