+++ app/api_v1/errors/exceptions.py
@@ -1,0 +1,10 @@
+class ValidationError(ValueError):
+    pass
+
+
+class AuthenticationError(ValueError):
+    pass
+
+
+class TokenExpiredError(AuthenticationError):
+    pass
