+++ app/api_v1/resources/ValueSet.py
@@ -1,5 +1,5 @@
 from app.api_v1.authentication import token_auth
-from app.api_v1.errors import *
+from app.api_v1.errors.errors import *
 from app.api_v1.rate_limit import rate_limit
 from app.api_v1.etag import etag
 from app.models.fhir.codesets import ValueSet
